When creating a new chunk, sync the penultimate.
This is to avoid an inconsistency in IDs, where the penultimate chunk
has some unsynchronised appends, and the new chunk has (effectively) a
synchronised "oldest". If the process dies before syncing the
penultimate chunk, there will be a discontinuity.

This fixes that by just syncing the then-final chunk before adding a new
one.